diff_stuff/tensorrt_inference.py
import tensorrt as trt
import pycuda.driver as cuda
import pycuda.autoinit
import numpy as np
from PIL import Image
import torch

class TensorRTDiffusionPipeline:
    def __init__(self, adapter_engine_path, unet_engine_path):
        self.logger = trt.Logger(trt.Logger.WARNING)
        
        # Load T2I-Adapter engine
        with open(adapter_engine_path, "rb") as f:
            runtime = trt.Runtime(self.logger)
            self.adapter_engine = runtime.deserialize_cuda_engine(f.read())
            
        # Load UNet engine
        with open(unet_engine_path, "rb") as f:
            self.unet_engine = runtime.deserialize_cuda_engine(f.read())
            
        # Create execution contexts
        self.adapter_context = self.adapter_engine.create_execution_context()
        self.unet_context = self.unet_engine.create_execution_context()
        
        # Allocate GPU buffers
        self._allocate_buffers()
        
    def _allocate_buffers(self):
        # Allocate buffers for T2I-Adapter
        # Input: 1x3x1024x1024 FP16 = 1*3*1024*1024*2 bytes
        adapter_input_size = 1 * 3 * 1024 * 1024 * 2
        # Output: Features from adapter (simplified)
        adapter_output_size = 1 * 3 * 128 * 128 * 2
        
        self.adapter_input_buffer = cuda.mem_alloc(adapter_input_size)
        self.adapter_output_buffer = cuda.mem_alloc(adapter_output_size)
        
        # Allocate buffers for UNet
        unet_sample_size = 2 * 4 * 128 * 128 * 2      # FP16
        unet_timestep_size = 1 * 2                    # FP16
        unet_encoder_size = 2 * 77 * 2048 * 2         # FP16
        unet_output_size = 2 * 4 * 128 * 128 * 2      # FP16
        
        self.unet_buffers = {
            'sample': cuda.mem_alloc(unet_sample_size),
            'timestep': cuda.mem_alloc(unet_timestep_size),
            'encoder': cuda.mem_alloc(unet_encoder_size),
            'output': cuda.mem_alloc(unet_output_size)
        }
        
    def preprocess_depth(self, depth_image_path):
        """Preprocess depth image for T2I-Adapter"""
        image = Image.open(depth_image_path).convert("RGB")
        image = image.resize((1024, 1024))
        image_array = np.array(image).astype(np.float32) / 127.5 - 1.0
        return np.transpose(image_array, (2, 0, 1))[np.newaxis, ...].astype(np.float16)
        
    def run_inference(self, depth_image_path, prompt="ancient statues along a river"):
        """Run full diffusion pipeline"""
        # 1. Preprocess depth image
        depth_input = self.preprocess_depth(depth_image_path)
        
        # 2. Run T2I-Adapter
        adapter_features = self._run_adapter(depth_input)
        
        # 3. Run UNet with adapter features (simplified)
        # In a real implementation, you'd need to:
        # - Encode prompt with text encoder
        # - Run scheduler for 4 steps
        # - Combine adapter features with UNet
        # - Decode with VAE
        
        print("Inference completed with TensorRT optimization")
        return "out_final_tensorrt.png"  # Replace with actual result
        
    def _run_adapter(self, depth_input):
        """Run T2I-Adapter inference"""
        # Copy input to GPU
        cuda.memcpy_htod(self.adapter_input_buffer, depth_input)
        
        # Set input shape
        self.adapter_context.set_binding_shape(0, depth_input.shape)
        
        # Execute inference
        bindings = [int(self.adapter_input_buffer), int(self.adapter_output_buffer)]
        self.adapter_context.execute_v2(bindings)
        
        # Copy output from GPU
        output_shape = (1, 3, 128, 128)
        output = np.empty(output_shape, dtype=np.float16)
        cuda.memcpy_dtoh(output, self.adapter_output_buffer)
        
        return output

# Usage example
if __name__ == "__main__":
    # Note: You need to run the export scripts and convert_to_tensorrt.bat first
    pipeline = TensorRTDiffusionPipeline("t2i_adapter.trt", "sdxl_unet.trt")
    result = pipeline.run_inference("out_depth.png")
    print(f"Result saved to: {result}")
