diff_stuff/requirements.txt
diffusers
controlnet-aux
huggingface-hub
safetensors
transformers
onnx
tensorrt
pycuda
numpy
Pillow